# docs/swagger.yaml
basePath: /
definitions:
...
  response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  response.Response:
    properties:
      body: {}
      status:
        type: integer
    type: object
  response.UserGenderSwag:
    enum:
    - Male
    - Female
    - Other
    type: string
    x-enum-varnames:
    - Male
    - Female
    - Other
...
  response.UserSwag:
    properties:
      avatar:
        type: string
      birthday:
        type: string
      description:
        type: string
      firstname:
        type: string
      gender:
        $ref: '#/definitions/response.UserGenderSwag'
      id:
        type: integer
      login:
        type: string
      middlename:
        type: string
      password:
        type: string
      phonenumber:
        type: string
      surname:
        type: string
    type: object
...
host: mailhub.su
info:
  contact: {}
  description: API server for MailHub
  title: API MailHub
  version: "1.0"
paths:
...
  /api/v1/user/avatar/delete:
    delete:
      consumes:
      - application/json
      description: Handles requests to delete user avatar.
      parameters:
      - description: CSRF Token
        in: header
        name: X-Csrf-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User avatar deleted successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Not authorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Delete user avatar
      tags:
      - users
  /api/v1/user/avatar/upload:
    post:
      consumes:
      - multipart/form-data
      description: Handles requests to upload user avatar.
      parameters:
      - description: CSRF Token
        in: header
        name: X-Csrf-Token
        required: true
        type: string
      - description: Avatar file to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: File uploaded and saved successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Error processing file or failed to get file
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Upload user avatar
      tags:
      - users
  /api/v1/user/count:
    get:
      consumes:
      - application/json
      description: Handles requests to get count user.
      parameters:
      - description: CSRF Token
        in: header
        name: X-Csrf-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Not authorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get count user
      tags:
      - users
  /api/v1/user/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Handles requests to delete user data.
      parameters:
      - description: CSRF Token
        in: header
        name: X-Csrf-Token
        required: true
        type: string
      - description: User ID to delete
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User data deleted successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Not authorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Delete user data
      tags:
      - users
  /api/v1/user/get:
    get:
      description: Retrieve the user associated with the current session
      parameters:
      - description: CSRF Token
        in: header
        name: X-Csrf-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User details
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Not Authorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get user by session
      tags:
      - users
  /api/v1/user/update:
    put:
      consumes:
      - application/json
      description: Handles requests to update user data.
      parameters:
      - description: CSRF Token
        in: header
        name: X-Csrf-Token
        required: true
        type: string
      - description: Updated user data
        in: body
        name: updatedUser
        required: true
        schema:
          $ref: '#/definitions/response.UserSwag'
      produces:
      - application/json
      responses:
        "200":
          description: User data updated successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Not authorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Update user data
      tags:
      - users
  /api/v1/verify-auth:
    get:
      description: Verify user authentication using sessions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Not Authorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: Verify user authentication
      tags:
      - users
swagger: "2.0"